refactor: optimize Context hashing, simplify predicate keys, and update derivation counting to use BigInt

+++ lib/src/parser/common/parse_result.dart
@@ -140,12 +140,12 @@ class MarkForest with Iterable<List<Mark>> {
 
   final LazyGlushList<Mark> inner;
 
-  int get derivationCount => inner.countDerivations();
+  BigInt get derivationCount => inner.countDerivations();
 
   @override
   Iterator<List<Mark>> get iterator => inner.allMarkPaths().iterator;
 
   Iterable<List<Mark>> allMarkPaths() => inner.allMarkPaths();
-  int countDerivations() => inner.countDerivations();
+  BigInt countDerivations() => inner.countDerivations();
   GlushList<Mark> evaluate() => inner.evaluate();
 }
